wbcalc,execgraph: Optimize WB calc for (hb U rf)-ordered writes
In cases where a program is "well-synchronized", writes are
ordered by (hb U rf_x). This commit takes advantage of this fact
and avoids calculating WB if all writes satisfy the above condition,
both in maximality checks and when checking for coherent rfs/revs.

The optimization is achieved via (1) a new 'ordered' map
(LOC->BOOL) which is checked every time we want to iterate the wb
successors or predecessors of a given store, and (2) actually
checking the stores one by one within a given view (for the
revisit case).

The optimization seems to help the WB driver come closer to the
MO one, although it does not make them even.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>