dep*,interp*,driver*: Redesign dep handling; restore Persevere
This commit redesigns how dependencies are passed between the EE
and the Driver so that we can restore Persevere.

System calls comprise multiple instructions; if one store of these
instructions initiates a recursive exploration in TruSt, then,
using the old dep handling mechanism, we would need to re-set the
dependencies of the next instruction after the recursive call is
over.

(This is not specific to Persevere, but rather is due to the
way deps are passed from EE -> Driver: the EE sets some pointers
which are then read by the Driver. Since recursive calls invalidate
these pointers we would have to re-set them before calling the
driver again.)

Re-setting dependencies that way is error-prone, however. Also,
for function calls that have e.g., data/addr deps, if we initiate
a recursive exploration during the function call, the set
dependencies are going to be lost and we will have no way of
recovering them.

This commit gets around the above issues by explicitly copying
and passing the dependencies to the Driver whenever the latter
is called. Since we copy the dependencies before calling the
Driver, we can actually also save them and keep them around.
(Although copying is a bit slower, it should not matter that
much: we iterate over these dependencies later anyway.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>