eventlabel,driver*,interp*: Redesign Driver-EE API
This commit completely redesigns the API between the Driver and
the Interpreter. The triggering event for this was the redesign
of the dependency-passing mechanism between the EE and the Driver.

Redeclaring all the Driver methods every time we want to change
a label or the way information is passed around is tedious (a lot
of parameters smell anyway), so we would either have to use
proxy objects (effectively simulating the EventLabel hierarchy)
or pass EventLabels directly to the Driver. We opt for the
second solution and thus make all necessary changes to accommodate
for the new API.

The implementation is about 2.5% slower on pinky, but I think
we can afford it.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>