aaccess,driver,interp,eventlabel: Introduce an AAccess type + fix sext bug
This commit introduces a type for atomic accesses packing
together the size and the type of a given access.

In turn, this new type is used in the interpreter and the
driver in order to correctly load values from memory when
necessary. (Until now some values were not properly sign-extended
if their size was less than 64 bits.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>