driver: Change the way access validity is checked
Instead of delegating the burden of tracking allocated addresses
to the Interpreter (which is complicated in the presence of
recursive subexplorations), simply introduce a more complex
access validity check. The new check is essentially the old
memory-race detection mechanism for dynamic addresses, as
if a malloc event is not found for the corresponding access then
we can simply deem it as invalid.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>