driver,execgraph: Graceful error reporting and quitting
This commit sets the foundations for more graceful error reporting
and quitting, by not calling exit() anymore when errors are
detected. Instead, it merely raises some flags which are taken
into account by the driver so that the latter quits the execution.

The commit also ensures that recursive subexplorations (stemming
either from replaying or from backward revisits) will be handled
properly and will not lead to memory corruption.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>