pass*,sexpr*,*: Ensure module information are collected properly
This commit makes sure that instruction metadata and annotations
are collected properly under TruSt.

Specifically, it ensures that metadata are collected relatively
early in the transformation phase so that we also collect info
for memcpys and memsets.

It also ensures that the annotation pass runs last so that the
module is stable and the info we get is for the version of the
module that is going to be used for the verification.

To reconcile the two requirements above, it also redesigns classes
related to annotations so that we can collect annotations in a
module-specific way, and then transform them into a module-agnostic
way, before passing the modified module back to the driver.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>