calc*,execgraph*,interp,*: Cleanup and persevere bug fix
Apart from some minor cleanups, this commit fixes a bug with
Persevere and concurrency, where the utilized file descriptors
would not be passed to other threads as part of the shared state.

A better solution would be to simply delete EE::fds and walk over
the graph to find the next available file descriptor.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>