imm,lkmm,cohcalc,*: View bug: don't keep porf views in dep models
In many cases, we need to treat holes specially in TruSt. For
instance, we do allow an fr edge from a deleted event A towards
the pporf-prefix of a revisiting write B, as long as A is a hole
and is also porf-before B.

To account for such cases, we would keep porf views in event
labels too. However, that turned out to be very wrong: when
a write revisits a read, it events in its porf prefix are being
deleted, and thus its porf view becomes invalidated.

This commit removes all porf-view calculations for dep-tracking
models, and instead checks for the existence of holes by simply
reading the pporf view of an event.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>