instannot: Fix soundness bug w/ load annotation
Before an annotation was passed to the driver, the annotation
was concretized: all uknown variables (but the one being
interpreted) were assigned their current values so that only
the load is unknown in the whole expression.

However, we did proprly exclude the load from the concretization:
assuming that the load's register had already a value (e.g.,
due to a previous iteration), it would have also been erroneously
concretized.

This commit takes care of this issue by (1) ensuring that the
instruction interpreted is not concretized, and (2) that there
is always one uknown in an expression that corresponds to the
annotated instruction (and not, e.g., to a subsequent extractinst
in the case of CASes).

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>