bitmask,eventattr,eventlabel,*: Refactor final + local flags for labels
This commit introduces an "attribute" field for writes that
subsumes the previous final and (the missing) local flags. It
is more extensible and much more maintainable than plain bool
flags in all different write subclasses.

As the newly introduced type WriteAttr is a bitmask, the
necessary infrastructure is also added to ensure type safety.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>