llvmmodule: Eliminate annotations after mem2reg + localsimplifycfg
In case a store is doubly annotated, the annotations and the store
might end up in different basic blocks, which messes up with the
annotations elimination mechanism (since the latter has to iterate
forward and find the respective store). An example where that
happens is ms-optimistic + tagging.

To ensure that we don't unnecessarily fail in such cases, make
sure that annotations will be eliminated after mem2reg and
localsimplifycfg have run.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>