mocalc: Cool refactoring
Instead of performing half the maximality checks for events of
the saved prefix (e.g., whether they are mo-after deleted events)
and half for the deleted events (e.g., whether they are rb-before
the saved prefix), perform all checks for one of the two
categories (in this case, deleted events), using a single function.

This is going to be helpful for the opt-marked optimization too.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>