assumeprop: Ensure that condition is always zexted when needed
assume() calls are not the only cases where the argument of a
propagated assume needs to be zero-extended. More generally,
this should happen when the argument of the assume function
has a typesize bigger than the typesize of the condition.
(The typesize cannot be smaller because the condition should
be i1.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>