immdriver,lkmmdriver: Fix bug when calculating hole pporf-views
Suppose we have a porf cycle (LB-style) ABCD, with AB \in T1 and
CD \in T2. Upon creating such a cycle (e.g., by having D revisiting
A), a non-atomic event E added between A and B should have B, C and
D in its pporf view.

Instead of that, however, the code would erroneously set E as the
max of its thread, leading to buggy behaviors in cna-c11-relaxed.

This commit addresses this issue.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>