lkmmdriver: Maintain ctrl deps toward RMW reads (WORKAROUND)
Consider the following thread in an execution where 42 is
the maximal value for x.

    a := x
    if (x == 0)
        assume(0)
    FAI(x, 1)

Assuming that the FAI only depends on a := x, the above thread
will look as follows when FAI(x, 1) backward-revisits some read:

    a := x
    EMPTY
    FAI(x, 1)

When the missing read of x is re-added it might read 0, and
the driver will subsequently try to add a BLOCK label in the
graph, though it will be unable to do so.

This can lead to all sorts of problems while replaying the graph
(as demonstrated by "qspin_pending_liveness.c"), so let's include
ctrl deps to CAS_R/FAI_R as a temporary workaround.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>