interp: Make sure that executeAtomicRMWOperation doesn't have UB
fetch_and_add() and relatives should have a 2-complement behavior
when dealing with signed integers. This commit ensures that these
functions will behave correctly by truncating and sign-extending
the result of the operation.

A better/complementary fix would be to ensure that loads simply
do not get a store's whole SVal but rather only read the bits
they care about, but the mixed-size accesses branch will take
care of that (and more). So let's deem this fix Good-Enough (TM)
and Deal-With-It-Later (TM^2).

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>