depinfo,deptracker,*: Change dep infrastructure
This commit changes how dependencies are passed from the EE to
the Driver and also adds the ability to store the dependencies
of an event into the respective label.

1. DepInfo is renamed to Deps
2. A new class DepInfo is created that is used to pass deps to
the Driver. If the Driver does not store the passed dependencies,
no copying of said dependencies is performed (in contrast to
the previous infra). No runtime benefit seems to occur.
3. Iterators for dependencies are added at EventLabel as well
as as wrappers in GraphIterators.hpp.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>