graphiter: SFINAE to avoid overloading copy ctor
The existing iterator infrastructure uses a template ctor. This
can shadow the copy ctor, and this is why we had to manually
overload it (both for LabelIterator and EventIterator).

With SFINAE, we can only enable the template ctor for classes
that are not derived from LabelIterator (i.e., just for
ExecutionGraphs and its derivatives).

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>