interp: Assume CAS load result is signed
Cf. unsigned-cas example.

Even though a CAS may be performed on a variable declared unsigned
(and there is no overflow), the compare operand of the CAS is a
signed integer in LLVM-IR.

Sign-extend the compare operand to ensure correct CAS behavior,
similarly to how FAI results are sign-extended.

(When mixed-size accesses are introduced, try to load/store/cmp
only chunks of values.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>