kater/relation.predicate,*: Better support for dom/codom
Until now, we would model a relation/predicate using a plain int.
Type-punning concerns aside, the int used to model a rel/pred
was also the index in a vector with information about the rel/pred;
this made modeling doms/codoms hard, as we would have to
explicitly use indices in the information vector to denote the
dom/codom of a relation.

To alleviate this issue, this commit introduces new Relation
and Predicate classes. It also introduces a PredicateSet class
to replace the horrible std::set<int>s for predicate collections.
Relation/Predicate has now a public enum with all builtins
available, and indices are no longer used implicitly as IDs
(e.g., in the SavedVarSet which is now a map Relation -> RE).
Finally, printing information specific to GenMC are moved into
Printer.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>