include,elimannots,*: Redesign annotation elimination mechanism
clang++ generates weird code when it comes to cmpxhg/atomic ops.
Instead of a simple block like the following:

     %res = cmpxchg
     jump %res ...

it can be the case that a switchinst on the CAS's mode is
performed, and then multiple cmpxchg instructions are generated:

      switch i32 %__m, label %monotonic [
          i32 1, label %acquire
          i32 5, label %seqcst
      ]
(where the respective labels contain the same CAS with a different mode).

In order to perform correct annotation elimination in such cases,
we introduce a __VERIFIER_annotation_end(int) primitive that matches
a previous __VERIFIER_annotation_begin(int) (i.e.,
dom(begin,end) + postdom(end,begin)), and all atomic instructions
in-between are annotated.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>