kater/visitor: Add a generic visitor class
It's best if we remove all the constraint solving functionality
from Constraint.cpp and move it to Kater instead (this is also
going to be helpful for printing declared relation names).

To that end, instead of implementing a classic visitor, make
a more unconventional visitor based on lambda overloading and
C++17 extensions. That way, we can have two different visitors
in Kater.cpp (one for checking inclusions and one for printing)
in a non-intrusive manner (Constraint.hpp does not have to be
changed).

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>