driver,tests: Ensure no revisit takes place during error replay
Suppose a store W triggers an access validity error in G,
and that G contains holes. In the metadata-collection phase
for the error, GenMC will attempt to fill the holes. However,
if some store W' filling some hole attemps to revisit, it
might try to examine the successors of W in co, thereby
leading to an out-of-bounds error (W is not in CO --- invalid
location).

This commit addresses this issue by forbiding revisits during
error replaying.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>