spinassume,tests: Consider more loops for spin-assume
Whether a loop will be consider for spin-assume depends on the
structure of its Φs. Assume the following IR:

entry:
    %l = load x
    %c = CAS x %l v
    %a = add %l 2
header:
    %f = Φ(%a, %l2)
    jc %f latch exit
latch:
    %l2 = load x
    j header
exit:
    ret 0

where a Φ is created because we cannot do the load and the
addition in the header, as the loaded value value is required
before the loop as well.

This commit makes spin-assume also consider these loops, by
allowing Φs the arguments of which come from loads or CAS compare
operands of the same variable.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>