execgraph,driver: Rework label-addition methods
Instead of (1) having to manually update label views every time
an event is added to the graph and (2) manually blocking a thread
whenever there is need, this commit introduces:

    - addLabelToGraph: adds a label and updates its views
    - blockThread: blocks a thread (by adding a label and signaling
      the interpreter)
    - blockThreadTryMoot: blocks a thread and checks whether the
      execution can be mooted
    - unblockThread

It also removes EG::add{Read,Write,Other}LabelToGraph() in favor of
EG::addLabelToGraph(). The latter is also more robust since it
replaces events with more care.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>