eventlabel,execgraph*: Stamps should be distributed by the graph only
They are just an intrusive way of keeping track of the addition
order. The graph should be the one determining.

This commit optionalizes stamps (it is OK for a stamp not to have
an assigned stamp; just don't use 0 for that), and makes the EG
assign them whenever it adds a label.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>