driver: Fix nasty bug showcasing in Z6.U
Z6.U would randomly crash. The crashes were attributed to
Kater and relation saving. This attribution, however, was wrong:
even with no saving and random testing the bug would still appear
from time to time.

It turns out that the problem was that, when revisiting RMW reads,
we would only check consistency for the read, and not for the
newly-added write.

This commit addresses this issue.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>