driver,execgraph*: Rework event addition/removal mechanism
remove() was only called with the last thread label as an
argument. Event addition would typically be performed at the
end of a thread, on top of some empty label, or (due to some
degenerate cases) on top of the last label of a thread.

This commit ensures that remove() is only called with the last
thread event, and that addition only happens at the end of a
thread or on top of a hole.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>