kater/*: Infrastructure for printing an hb getter (FIXME)
This commit adds some infrastructure so that the Printer can
print a getter for a user-specified HB. That HB needs to be
stored in a view.

As HB is only necessary for some error checks, this is a
temporary commit until all error checks are performed by
Kater. (The latter requires a rewrite of the Predicate
representation.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>