driver*,kater/kat: Fix hb-view bug in RC11/IMM (INTERESTING)
After HEAD~2 fixed a bug in hb-code printing, some test cases
started yielding the wrong number of executions. Many tests in
IMM were broken because hb was not properly calculated for that
model.

However, many tests were also "broken" for RC11 as well, but
in this case due to the different definition of release sequences.
Interestingly enough, chase-lev yields a different number of
executions depending on the relseq definition used.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>