eventlabel,driver: Fix concurrency bug
When implementing instruction caching, we moved ThreadInfo
from ThreadStartLabel to ThreadCreateLabel (as otherwise
handleCreate/visitCreate would have to take two arguments,
which would break the dispatching switch).

The problem with the above change is that, when creating
an execution context, the parent create event might not
exist, as a result of which creation will fail.

This commit fixes this issue with duplication: now ThreadInfo
is a member of both ThreadCreate and ThreadStart.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>