driver: Fix (concurrency) issue with instruction caching
Under dependency-tracking models it can be the case that
we try to cache a label without all of its predecessors
in the existing cache entry. (This can be the case e.g., when
revisiting some filled hole and trying to cache the RMW
successor --- see cinc(4).)

This commit rewrites the caching mechanism so that it also
works when some po-predecessors of a given label have not
been cached.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>