driver: Fix a last concurrency issue
1. It is not safe to pop the current execution state after
submitting a task to the global workqueue: after revisiting
an RMW read, its corresponding RMW write might be added, which
in turn might add some revisit items in the local workqueue.
(Popping will destroy that queue.)

2. If we are to keep the current state around, it is not
safe to move the execution graph (nor the allocator or
the cache).

The last point implies that we might make some unnecessary
graph copies, but that should be fine.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>