include,interp*: Split barrier_wait() into two operations (bugfix)
Each interpreter instruction should map to at most two driver
operations. barrier_wait() was mapping to three, the last
read operation would sometimes be replaced with a block label,
and this would create problems when replaying.

(The interpreter is not supposed to replay blocked instructions, but
it had no way of only replaying the first two barrier labels and not
the third.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>