driver,tests: Fixes for BAM and error replaying (bugfix)
Suppose an error is detected at a graph with some barrier_wait()
operations unfinished (modeled as FAI+B). When replaying, B
is removed from the replaying view, but when the EE reaches the
corresponding event (a BWaitRead), it will exist in the execution
graph as a blocklabel, causing all sorts of issues (see parker/2).

This commit fixes the issue by changing the blocking representation
to FAI+R+B so that replaying will work properly.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>