driver: Fix CAS IPR bug
When checking whether a CasRead blocks reading maximally, we
cannot simply gets the read value, as the access may be invalid.
This commit adds the necessary guards to ensure that values
of invalid accesses are not fetched. (If the access is indeed
invalid, either an error will be reported later on, or the
graph is inconsistent, and the read will be revisited.)

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>