interp,driver: Fix uninitialized memory checks (squashed)
These checks were happening all over the place. This commit
introduces checkInitializedMem(), which ensures that reads/writes
access properly initialized memory. The interface is also now
uniform with checkAccessValidity() and checkForRaces().

To that end, it also introduces the ErrorDetails class in order to
reduce the number of parameters passed to reportError.
Introducing this class also allows us to construct and return
ErrorDetails objects which can be then printed by means of
reportError.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>