eventlabel: Track down UB (aka: "should I rewrite it in Rust?")
Generating new consistency checks from Kater would lead to
GenMC sometimes reporting more executions for RWC+syncs. The
new consistency checks were successor-based (the problem did
not exist before), and the bug would only manifest in the
CI/CD servers: the only time it manifested locally, a different
(but similar) test would break, namely W+RWC.

Of course, this all smelled like UB to, but was quite difficult
to track down, as it wasn't locally reproducible. What worked
in the end was

(1) getting the generated consistency checks from the server as
part of the output (different memory layout and non-deterministic
iteration in certain kater contains meant that the consistency
checkers generated there were different than the ones generated
locally)

(2) building with the same LLVM version as the server, and
running the test with the exact same CLI options, even if some
of them (e.g., -disable-estimation), clearly shouldn't affect
the test result.

The above sufficed to reproduce the bug with a fair frequency.
Then, sanitizing the code showed that the uninitialized
ReadClass::ipr field is the likely cause of the bug. (It's kind
of surprising that this didn't trigger UB in the past,
but oh well...)

This commit also initializes some class members that were not
necessary to initialize, but as the old adage goes
"whoever gets burned with the mush blows at the yoghurt too" :-)

NB: It turned out that kater-generated checks were the culprit:
there was an uninitialized field on the printed drivers. The
ipr field just made the issue manifest more often (and helped
in debugging quite a bit).

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>