saddr,*: Change how addresses are modeled
This commit changes the address representation so that it includes
the allocating thread. By doing that we (1) do not have to save
allocation events in our cache (even though they are still included
now), and (2) do not have to have the cache as part of the
driver state and pass it to other worker threads (this was
necessary due to subtle concurrency issues; see tool paper for
an example).

A relevant test showcasing the potential benefits in performance
is included, though it's unclear how much this helps in
non-pathological scenarios.

This change is performed to help fuzzing heuristics.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>