execgraph,*: Move getRevisitable() to the driver
There is no point in getRevisitable() being a member function
in EG. There is also no point in getRevisitableFrom() being
in the consistency checking path, is this function used to
incorporate optimizations specific to maximal extensions,
while consistency checking can be used in other parts as well.

This commit changes the consistency checking API so that it
filters out non-coherent stores (instead of collecting+returning),
and moves the collection of revisitable reads to the driver.

Signed-off-by: Michalis Kokologiannakis <mixaskok@gmail.com>