execgraph,*: Do not keep orphaned begins in graph
We generally keep begin events in the graph so that TIDs do not
lose their meaning (e.g., after revisiting). However, this could
lead to graphs with "orphan begins" (e.g., if, upon revisiting,
the corresponding create was never created).

This was not an issue per se, but we had to be careful when starting
a new execution (handleExecutionStart) not to create stack frames
for orphaned threads. As this makes integration with other
runtimes more difficult, this commit removes orphaned begins.