promotepass,*: Allow src/dst to have different (yet compatible) types
When doing memcpy(dst,src,len), we used to always cast the src
type to dst. Oftentimes, however, it is also useful to allow the
cast the other way around (e.g., when dst is the result of an
allocation).

This commit enables such promotions, and also adds support for
fortified memcpy (often emitted in macOS).