driver: Allow printGraph() to be called when EE is not set
Graph printing was relying on calling methods exposed by the runtime.

As this it not the case for all runtimes, this commit enables graph
printing even when EE is not set. The only caveat is that no metadata
can be printed in the case, and hence printGraph() has to be called
with printMetadata=false.